backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.0.1
pymongo==4.5.0
google-generativeai==0.3.2
python-multipart==0.0.6
email-validator==2.1.0
pypdf==5.3.1
python-docx==1.1.2
Flask-SocketIO==5.5.1